f5/bigip: move the Rust parser bridge into the product tree
Relocate the canonical (de)serialisation bridge from tests/ to
dialects/f5/bigip/parser/_rust_bridge.py so the f5 command (and every
other BigipConfig consumer) can reconstruct the typed dataclasses from
the Rust parser's JSON in production — not just in tests. The parity
test now imports the bridge from its product location.

Import-linter contracts stay green (7 kept); ruff clean; round-trip
identity test still passes across the corpus.
